# application/services/custom_services/featureinventory/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: gepinventorys open api specification
   description: featureinventory API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8011
     description: featureinventory
tags:
- name: featureinventory
paths:
   "/featureinventory/{id}":
      delete:
         tags:
         -  featureinventory
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  featureinventory
         description: GpGetEntityById
         parameters:
         -  in: "path"
            description: GpGetEntityById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/featureinventory/get/search":
      get:
         tags:
         -  featureinventory
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/featureinventory/get/update":
      put:
         tags:
         -  featureinventory
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/featureinventory"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/featureinventory":
      put:
         tags:
         -  featureinventory
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/featureinventory"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  featureinventory
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  featureinventory
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/featureinventory"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/featureinventory/userid/created_by":
      get:
         tags:
         -  featureinventory
         description: GpGetNounCreatedBy
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      featureinventory:
         type: object
         properties:
            item_uuid:
               type: "string"
               description: Description
            name:
               type: "string"
               description: Description
            type:
               type: "string"
               description: Description
            description:
               type: "string"
               description: Description
            attachments:
               type: "string"
               description: Description
            price:
               type: "string"
               description: Description
            price_currency_type:
               type: "string"
               description: Description
            cost:
               type: "string"
               description: Description
            cost_currency_type:
               type: "string"
               description: Description
            tag:
               type: "string"
               description: Description
            isMatrix:
               type: "string"
               description: Description
            matrixid:
               type: "string"
               description: Description
            warehouse:
               type: "string"
               description: Description
            warehouse_loc:
               type: "string"
               description: Description
            lot_or_batch_num:
               type: "string"
               description: Description
            manufacturer:
               type: "string"
               description: Description
            manufacture_country:
               type: "string"
               description: Description
            external_reference_one:
               type: "string"
               description: Description
            gephistoryid:
               type: "string"
               description: Description